Fix rounding error in stellar age computation that can lead to ever-so-slightly negative ages just after birth.

+++ src/feedback/EAGLE/feedback.h
@@ -102,7 +102,7 @@ INLINE static double feedback_get_enrichment_timestep(
     return cosmology_get_delta_time_from_scale_factors(
         cosmo, (double)sp->last_enrichment_time, cosmo->a);
   } else {
-    return time - sp->last_enrichment_time;
+    return time - (double)sp->last_enrichment_time;
   }
 }
